test: isolate stack route oracle tests
